+++ CCGW_Reborn/GameObject.h
@@ -25,7 +25,7 @@ public:
 	virtual void render(const GLuint &programID, const glm::mat4 &viewMat);
 	glm::vec3 getPosition() const;
 	void setPosition( glm::vec3 position );
-	void setScale(int scale);
+	void setScale(float scale);
 
 	GameObject(const GameObject& ref);
 	GameObject(glm::vec3 position, float scale);
